requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==2.0.1
pillow==10.4.0
tqdm==4.66.4
pyyaml==6.0.2
albumentations==1.4.8
opencv-python==4.10.0.84
matplotlib==3.9.0
